requirements.txt
opencv-contrib-python
numpy
matplotlib
keras
tensorflow
python-socketio
Flask
eventlet